Add files via upload
A program that calculates the salary for the week of the employees listed below. If an employee worked more than 40 hours that week, then calculate the remaining hours as time and a half (1.5x the normal rate). The program should print their name and salary earned. This program has a data structure which is List and a nested List, a user-defined function and a loop and function calling under it.